Split shader into its own module
